+++ java1/src/Vertex.java
@@ -9,10 +9,10 @@ public class Vertex implements java.io.Serializable, Comparable<Vertex> {
 	final static long serialVersionUID = (long) "Vertex".hashCode();
 
 	// coordinates of the vertex in Matlab's order (y, x)
-	private final float[] coords = new float[2];
+	private final double[] coords = new double[2];
 	
 	/** The standard constructor for class Vertex. */
-	public Vertex(float y, float x) {
+	public Vertex(double y, double x) {
 		coords[0] = y; 
 		coords[1] = x;
 	}
@@ -24,20 +24,20 @@ public class Vertex implements java.io.Serializable, Comparable<Vertex> {
 	}
 	
 	/** Get the coordinates of this Vertex. */
-	public float[] coords() {
+	public double[] coords() {
 		return coords;
 	}
 	/** Get the coordinates of an array of Vertex objects. */
-	public static float[][] coords(Vertex[] v) {
+	public static double[][] coords(Vertex[] v) {
 		if (v == null) return null;
-		float[][] out = new float[v.length][2];
+		double[][] out = new double[v.length][2];
 		for (int i = 0; i < v.length; i++)
 			out[i] = v[i].coords;
 		return out;
 	}
 	
 	/** Create an array of new Vertex objects from a coordinate array. */
-	public static Vertex[] createVertices(float[][] coords) {
+	public static Vertex[] createVertices(double[][] coords) {
 		Vertex[] out = new Vertex[coords.length];
 		for (int i = 0; i < coords.length; i++)
 			out[i] = new Vertex(coords[i][0], coords[i][1]);
@@ -45,20 +45,20 @@ public class Vertex implements java.io.Serializable, Comparable<Vertex> {
 	}
 	
 	/** Move the Vertex to a new location. */
-	public void move(float[] newcoords) {
+	public void move(double[] newcoords) {
 		if (newcoords.length != 2) return;
 		coords[0] = newcoords[0];
 		coords[1] = newcoords[1];
 	} 
 	/** Translate the position of the Vertex by delta. */
-	public void translate(float[] delta) {
+	public void translate(double[] delta) {
 		if (delta.length != 2) return;
 		coords[0] += delta[0];
 		coords[1] += delta[1];
 	}
 	
 	/** Draw a line connecting the two Vertices v and w. */
-	public static void drawConnection(float[][] image, Vertex v, Vertex w) {
+	public static void drawConnection(double[][] image, Vertex v, Vertex w) {
 		Misc.drawLine(image, v.coords, w.coords);
 	}
